Added bmin/bmax and units as traits as well
Added custom traits type to allow None to still be used, while also
allowing it to be edited through a TextEditor. No other traits should
need to be added at this point from the point of view of a user
modifiable traitsui. Unless I misunderstand the use of ext_bounds and
ext_force_positive.