Iterators for Signal(.data), suppressing connections
Implements two iterators for iterating through the navigation axes of a
Signal, and returning either a sub-Signal, or a slice of the data array.
As it works with a copy of the AxesManager, it does not trigger any
indexing connections on the original one, which can improve speed for
those applications where those triggers are unnecceary.