Added bounds check flag
If changing left,right from 4,7 to 10,13, then previously you would have
to set right first, or left would just keep its old value (as it was
greater than right). If going the other way, then left would need to be
set first. To avoid having to check this when setting both, an internal
flag is added, so that the bounds cheking can be disabled. Note: Use the
flag responsibly!