Changed HE signal_data_function to lazy
Previously, the signal_data_function in HyperExplorer was only used in
plot(), where it was copied to the SpectrumLine/Image. Changing it
afterwards had no effect. With this change, the spectrumLine/Image will
always use the signal_data_function on HyperExplorer. This could maybe
be achieved with lambda, but I felt def was clearer. Should maybe be
done for navigation_data_function as well.