NEW: ModifiableSpanSelector.can_switch
Added an attribute and accompanying methods on ModifiableSpanSelector to
allow left and right edge modification to switch if the mouse tries to
cross the edge. The left value will still always be the smallest, simply
which edge the mouse is interacting with will change. the default value
for the attribute is False, which will correspond to the old behavior.
However, DraggableResizableRange will set it to True for it's use, to
correspond to the similar behavor for other widgets in the same
hierarchy.