NEW: Interactive ROIs
This commit establishes the connection between ROIs and widgets by
adding a new class BaseInteractiveROI, which contains the interface for
having a ROI have a corresponding widget. It also tries to supply a
reasonable default implementation of this interface, but several
functions will need to be overridden (raises NotImplementedError if
not).