ENH: Copy sets for iterations for safety
If any of the callbacks for an event modifies the connections, the
iteration would previously raise an RuntimeException (and rightfully
so). By copying the sets before iterating, we protect against this. An
example case would be a close or remove event, whose callbacks would
also want disconnect from that event.