Added option to reuse fig in estimate_image_shift
Previously, if `plot` was `True`, a figure would be created for each
alignment step. This commit gives the option to pass an existing figure
as `plot`, which well then be receive the alignment plots. This
effectively allows for the reuse of a figure, so that the alignment
process can be monitored live, but without creating more than one
figure.